feat: 유저의 하루 모임글 작성 개수 제한을 반환 타입을 long에서 int로 변경
- 처음엔 제한이 바뀌거나, 할 수 있어 long으로 했지만 개수 제한이 그렇게 많아지지 않을 것 같아 int로 변경

+++ src/main/java/com/momo/meeting/persist/repository/MeetingRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface MeetingRepository extends JpaRepository<Meeting, Long> {
 
-  long countByUser_IdAndCreatedAtBetween(Long userId, LocalDateTime startOfDay, LocalDateTime endOfDay);
+  int countByUser_IdAndCreatedAtBetween(Long userId, LocalDateTime startOfDay, LocalDateTime endOfDay);
 }
